Migrate the tick counter (uwTick) to a uint64_t to avoid clock overflow.

+++ STM32Cube_FW_H7/Drivers/STM32H7xx_HAL_Driver/Src/stm32h7xx_hal.c
@@ -68,7 +68,7 @@
 /** @defgroup HAL_Exported_Variables HAL Exported Variables
   * @{
   */
-__IO uint32_t uwTick;
+__IO uint64_t uwTick;
 uint32_t uwTickPrio   = (1UL << __NVIC_PRIO_BITS); /* Invalid PRIO */
 HAL_TickFreqTypeDef uwTickFreq = HAL_TICK_FREQ_DEFAULT;  /* 1KHz */
 /**
