+++ STM32Cube_FW_L4/Drivers/STM32L4xx_HAL_Driver/Inc/stm32l4xx_hal.h
@@ -615,7 +615,7 @@ typedef enum
 /** @addtogroup HAL_Exported_Variables
   * @{
   */
-extern __IO uint32_t uwTick;
+extern __IO uint64_t uwTick;
 extern uint32_t uwTickPrio;
 extern HAL_TickFreqTypeDef uwTickFreq;
 /**
